requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crewai==0.10.0
curl_cffi==0.6.0b9
dataclasses-json==0.6.4
Deprecated==1.2.14
distro==1.9.0
duckduckgo_search==4.4.3
Flask==3.0.3
frozenlist==1.4.1
googleapis-common-protos==1.62.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.2.post1
langsmith==0.0.87
lxml==5.1.0
MarkupSafe==3.0.2
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pandas==2.2.3
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-frontmatter==1.1.0
pytz==2025.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.26
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.5.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2025.1
urllib3==2.2.0
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
